feat(ui-kit): rename AdminCourseFilterModal, adjust clear filters functionality to remove also the grid's colum filters
