+++ packages/translations/src/lib/dictionaries/en.ts
@@ -31,7 +31,7 @@ export const EN: TDictionary = {
       hideActions: 'Hide actions',
       selected: 'Selected',
       sendNotification: 'Send notification',
-      clearfilters: 'Clear filters',
+      clearFilters: 'Clear filters',
     },
     courseCard: {
       courseEmptyState: {
